Fix race condition in rack-attack oAuth token specs
When running the Rack::Attack request spec shared examples, it may be
possible for a race condition to occur. To check that Rack::Attack will
"unlock" a token after the specified throttling period, the specs use
`travel_to(period)` to change the Ruby time into the future. However,
one of the oAuth tokens generated for the spec is set to expire in
`period + 1` . If the clock ticks over to the next second while making
the unlocked-period checks, then we would get an expired token error
instead of the expected response.