Fix RSpec/BeforeAll offenses
Also restructures the shared_example: "security findings finder" to
allow hooking into the "when the pipeline has security findings"
examples.

This was necessary as there are slight differences between
`before(:all)` and `before_all`. It appears the old implementation
relied on a quirk of `before(:all)` that allowed any specs after the
line where `before(:all)` was called to use the records created during
the `before(:all)` hook.

Co-authored-by: MOHAN SAI KRISHNA MOPADA <mohansaikrishnamopada2003@gmail.com>