Display unverified badge for signed commits without a signature
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/366364

**Problem**

When a commit signed with a X.509 certificate that has an error, we
don't display the signature status at all.

It happens, because we don't store invalid X.509 certificates and
`signature` value is `nil` for them.

**Solution**

If the commit has some signature but missing a certificate, we should
still display an `Unverified` badge for these commits.

Changelog: fixed