Ignore manually created MergeRequestsClosinIssues records
We will now allow setting a work item to be closed when an MR is
merged, from the work item side. For this reason, we need to make
sure these records don't get deleted whenever the merge request
description is updated.

We update manaully created records only if the new reference
in the MR description matches an existing manually created record.

We also need to rename the closes_work_item column to something
more appropriate.

Changelog: other