Add dependency aggregations finder
This finder implements a more performance version of the query used for
`Group#sbom_occurrences`. Is able to perform aggregations one page at a
time, while the current implementation of `Group#sbom_occurrences` must
read all the records in the group before aggregating them and then
finally applying the limit.