Remove synced_work_item param
Removes the synced_work_item param from `WorkItems::UpdateService` and
`WorkItems::CreateService`. This param was necessary when we created the
work item from the epic services through the Create/Update service of
the work item. With the param we for example skipped creation of notes
on the Work Item side.

We now directly use ActiveRecord to create a work item in the
Epic service and therefore no longer need to skip anything in the
WorkItem services.