Update type of subscription_add_on_purchases started_at column
Updates the started_at column in the subscription_add_on_purchases
table. That column was originally created with precision to match
the created_at that was used before the started_at column was
created, however the data being sent from the customers portal
application only provides date level precision. So the column should
be updated to match the granularity of the data it will actually
hold.

Changelog: added
EE: true