Set parent for legacy epic on work item creation
When a parent gets set via the hierarchy widget when creating an epic
work item, we can't rely on the WorkItems::ParentLinks::CreateService
to set the `parent_id` and `relative_position` because the Epic record
does not exist at that time.

Instead of creating the epic record before the work item, we now set the
`parent_id` and `relative_position` based on the ParentLink of the work
item on creation.