Drop redudnant vuln_reads indices, improve unnested_in_filters
The 2 indices dropped in this MR are redundant to two similar indices
which include the `dismissal_reason` column. However the implementation
of unnested in filters would not have worked with these indices as it
would include columns not accounted for. Thankfully it's position at
the end of the index means the optimisation can work exactly as it does
with some small modifications.

Changelog: changed