Use label text color for remove button icon in MR filter bar token
This addresses a dark mode style bug. Currently, the `x` button to
remove a label filter token in the MR filter bar is either the standard
token remove color or an inverted color for labels that have a darker
background color. This works well in light mode. But in dark mode, the
sass variables used are the inverse of what is expected, and the `x`
button ends up having an unexpected color.

This MR simply uses the label text color that is applied to the text of
the token for the remove button, so that they will always match for
label tokens.

Changelog: fixed