Remove error in Sidekiq routing docs
The comment and line are supposed to demonstrate how to "route all
import workers to queues generated by the worker name". But it's routing
to a specifically named queue instead.

(The feature_category should also be importers, and not import).

The example  was changed in e1188d860b1c3a4a5bab6b99b792d6a3d7f9c186
from routing to `nil`` to routing to `'import'` due to this suggestion
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/103228#note_1166851686

This change removes the example, as according to that suggestion we
don't want to encourage routing to `nil`, and so the example serves no
purpose.