Fix Debian package flaky specs
We have specs that are failing because the spec has assumptions on the
order of the files returned by `incoming_package.package_files`.

This MR fixes the specs by searching for the package files by filename,
instead of assuming the source files is the first file, the dsc file is
the second file, etc.