Generic Registry: Fix a bug when downloading files with sign symbol
We recently supported directory structure in generic package registry. This means we now accept uploading/downloading files that contain '/' in their names. To do this, we are encoding the filenames before saving them in the database. However, this caused an issue when the filenames has a + symbol. Because encoding a string that includes a + symbol would convert it into '%2B', and we don't want to do this. We just need to encode the filenames that include '/'.

Changelog: fixed