Import fork MR when target branch couldn't be created
During merge requests import when we import a fork
merge request that does not have target branch in the repo
there can be a situation where exported mr does not have target_branch_sha,
in which case target branch fails to be created and the MR fails
to be created as well.

Similar to what we do with source_branch, ignore target_branch creation
failure to get MR to be created so it is viewable and not completely lost.

Changelog: fixed