Set the start of the ID sequence for `build_stubbed`
Projects with the same ID refer to the same repository on the disk and
the IDs of the records initialized by the `FactoryBot.build_stubbed`
method can conflict with the ones created by the `FactoryBot.create`
method.

For this reason, we need to set the initial value for the ID sequence
used for objects initialized by the `FactoryBot.build_stubbed` method
high enough to prevent conflicts.