List all users in the Admin > Users
In the **Admin > Users** search, list all users (except ghosts) if no
filter is applied. Before, it only listed active users. This made the
search very difficult if the admin only knew the name/email but not the
state of the user. This bug was an overlook in a recent replacement of
the navigation by a filtered search:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150288

In addition, we added a new `Active` filter to look for only active
users.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/461672

Changelog: fixed