Adds support for dynamic schemas with date range filters
Adds a fallback for dynamic schemas so if a table doesn't exist in
the filter map, it will fallback to the `date` field. This means that
all dynamic schemas should use this field for them to function. We
control the dynamic schemas structure, so that shouldn't be a problem.

Changelog: added
EE: true
Signed-off-by: Robert Hunt <rhunt@gitlab.com>