Fix updated_at mismatch when removing parent link
When removing the WorkItems::ParentLink and we sync it to the epic, we
got a mismatch on `updated_at` for the work item and the epic.

The reason is that on the work item side we remove the ParentLink
record, while on the epic side we modify the `parent_id` on the epic
record.

As a fix, we keep the existing behaviour where changing the parent from
the epic side updates the `updated_at` on the epic and the work item,
while from the work item side, we do not update the `updated_at` value.