Fix flaky test that verifies secret push timeout[skip secret detection]
When RSpec's main process terminates and attempts to clean up
child processes upon completion, it terminates subprocesses
where the scans might be still ongoing. This behavior is not
recognized by the upstream library (parallel), which manages
all forked subprocesses it created for running scans.
When the upstream library attempts to close its forked
subprocesses which already terminated, it raises an
'ArgumentError' with the message 'bad signal type NilClass,'
resulting in flaky failures in the test expectations.