db: Document shards table as being cluster-wide
Strictly it's not needed for the shard names to be unique across the
whole cluster, but it's safer and easier to do so. This table won't be
very big anyway, maybe a couple hundred of lines.

This also set the feature category to Source Code Management. Gitaly
isn't involved in how Cells operate. Cells are managed on GitLab Rails'
side, and the Gitaly team doesn't know much about that.