Create parent links for imported epics
When importing epics from groups where epics and sub-epics features
where enabled to groups where these features are not enabled, we need to
retain the relationship on the work item side on imports.

We store the `parent_id` on the epic record. For this reason we bypassed
the license checks in the epic services and still created the sub-epic
relationships.

This doesn't work on the work item side, so we need to explicitly create
the work item parent link when importing epics.

Changelog: changed