Refactor required stops and avoiding it docs
We have many reasons to add a required stop. Database is just one of
them. This change moves the instructions on how to add a required stops
to the avoiding required stops page, instead of the database targeted
required stops page.

Additionally, it adds crosslinking from the database page to the general
information page.

Finally, it also adds clarification on how to plan for a required stop.