Update StuckExportJobsWorker to not fail parallel export jobs
Update StuckExportJobsWorker to not mark ongoing parallel
export jobs as failed. The job is completed and the record
is marked as started because there are parallel relation
exports still happening, so we don't need to mark the parent
project_export_job record as failed until the 6 hour expiration
mark hits.

Changelog: fixed