Improve GraphQL analyzer timeout handling
In the last iteration,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154466, we made
sure that there would be data present in case of an error. But this
resulted in an error object getting saved into `results[:complexity]`,
which makes it harder than necessary to identify logs with timeout
errors. For example, in our ElasticSearch instance, this causes the
`complexity` to get ignored.

See https://gitlab.com/gitlab-org/gitlab/-/issues/462387