Fix 500 error for /api/v4/groups?all_available=false&statistics=true API
The `GET /api/v4/groups?all_available=false&statistics=true` API was
throwing internal server error when logged in using an admin user
because the `SUM` aggregate was used to calculate the sum and all
the columns from the subquery did not appear in the `GROUP BY` clause
which resulted in a semantically incorrect query.

Here we have fixed the issue by using a window function to calculate the
sum of a column for a given `namespace_id`.

We change the query from:

```sql
SELECT namespaces.*,
       SUM(project_statistics.storage_size)
FROM namespaces
         LEFT JOIN project_statistics ON namespaces.id = project_statistics.namespace_id
GROUP BY namespaces.id;
```

to:

```sql
SELECT namespaces.*,
       SUM(project_statistics.storage_size) OVER (PARTITION BY project_statistics.namespace_id)
FROM namespaces
         LEFT JOIN project_statistics ON namespaces.id = project_statistics.namespace_id;
```