Fix serialization error on CI Lint with including a CI component
The CI Lint response includes the included files of the CI Config.
This included files response returns metadata of each included file.
The problem is the component inclusion because its metadata contains
"component_attrs" which includes a Project object.

The error occurs because serializing a Project object is forbidden.

This change fixes this problem by specifying the needed fields of
included files' metadata. We don't need to serialize "component_attrs".

Changelog: fixed