gitaly: Pass through input block of gitaly_client_call
`GitalyClient.call` method supports passing a block. This block yields
the arguments of RPC call to Gitaly. It allows the caller to modify
the arguments such as metadata. This method is never called directly. It
is wrapped by `gitaly_client_call`. Unfortunately, the wrapped method
doesn't pass through the input block. This commit fixes this bug so that
we can start adding custom RPC metadata to a specific RPC call.