Remove clean_merge_jid call
To reduce the number of SQL update queries during merge, we now
set the `merge_jid = nil` when transitioning `MergeRequest#state`
to `merged`. This will result to a single UPDATE query to set
`state_id` and `merge_jid` instead of having a separate UPDATE
query for setting `merge_jid = nil`.