Add a banner informing about token expiration
Display a banner to all users, similar to broadcast banners, for GitLab
version between 16.0 and 17.0, both included.

The banner informs about upcoming expiration of tokens and how to
troubleshoot the expiration.

The banner is dismissabled. The data is persisted in the browser cookie
and expires in one year from the date it has been dismissed. Like other
broadcast banners, if the user opens an incognito window the banner
reappears.

Changelog: changed