Add capability to handle labels unified between epic and work item
While migrating Epic to Work Item for a short period of time we
need to be able to handle some associations, in this case lables,
linked to both the Epic and its correspinding Work Item.

This would allow us a smooth transition from using the Epic object
to using its correspondent Work Item.