Upgrade semver_dialects to v3.0.1
This further improves the accuracy of Maven version parsing so that it
correctly matches against advisory and license ranges. This is important
to do especially in the context of License Scanning because it triggers
the ReactiveCachingWorker to keep the results up to date. This results
in a lot of errors reported for valid versions that are incorrectly
reported as invalid.