Use primary DB for authenticating artifacts downloads
A CI job downloads job artifacts from the `/api/v4/jobs/:id/artifacts`
endpoint. Previously the endpoint used any replica to authenticate the
current user via a job token, but that token depends on the job record
being in the database. However, there are no guarantees that the
replica has an up-to-date record of that job. As a result, users could
see intermittent 401 errors due to replication lag.

To avoid this, use the primary database when authenticating the build.
This commit adds a `ci_job_artifacts_use_primary_to_authenticate`
feature flag to roll this out.

Note that the runner API attempts to select an up-to-date replica for
the job that produced the artifacts, but it has no good way of
determining the job ID that originated the request for downloading
artifacts. In addition, the user authentication happens before the
replica selection happens.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/466138

Changelog: fixed