Refactor NPM metadata endpoint
This is currently a concern shared among the project, group, and
instance endpoints. But these three have diverged, and as a result we
have a number of `if endpoint_scope` checks in the endpoint code.

This refactor splits off the project metadata code and removes the `if
endpoint_scope == project` checks.