Respect carraige return in new job log viewer
This fixes a bug in the new job log viewer where the carriage return
character is ignored in lines, which causes all outputs of progress
messages (e.g. the git “Counting objects” message) to be rendered in a
single line.

This makes the viewer respect the carriage return and only show the
latest variation of the message in the log.

This change is also compatible with Windows runners that use CRLF.

Changelog: fixed