Remove unneeded Capybara::Session.new
When we use `using_session`, Capybara will already handle the session
for us, and we don't need to manually create the sessions.

Beside, it's the wrong way to use it. That's not how we should use
the sessions.

Ideally, we should use a fixed number of sessions. The default session
is called `default` and it's reused for all specs. The other sessions
should be reused for all specs as well, thus they should use the same
names.

We'll want to have a standard way to use the other sessions. Here
we're just removing the bogus usages.