Add Google Cloud IAM integration metrics to Service Ping
Many metrics exist for integrations, but right now we are not tracking
them for Google Cloud IAM integration because we lack the metric
definition files for them. This adds definition files for some of the
GCP IAM integration metrics, to help us track the success of this
feature.

Changelog: changed
EE: true