Improve TOC functionality in rich text editor
Previously, table of contents in rich text editor would
add empty bullets for missing headings, for example if a
h2 is present without an h1, it would add double bullets.
Now, only one bullet is added since an h2 without an h1
is considered at the same level as an h1.

Changelog: changed