Check deploy_key exists during ProtectedRefDeployKeyAccess#valid?
Prior to this change we did not validate that the `deploy_key` actually
existed when creating a `ProtectedRefDeployKeyAccess` object.

We have foreign_key constraints but this means we relied on the database
layer to raise exceptions instead of returning user friendly errors.

This change also modifies how we check if a
`ProtectedRefDeployKeyAccess` is a `deploy_key` record. Previously we
would check if deploy_key existed for
`ProtectedRefDeployKeyAccess#type => :deploy_key`.  This meant that
passing an invalid `deploy_key_id` would result in
`ProtectedRefDeployKeyAccess#type => :role` when it should be considered
a `deploy_key` type access level.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/461213

Changelog: changed