Refactor graphql instrumentation tracer
The GraphQL ruby tracers don't work as we expected them to. An
`#execute_query` trace does not necessarily execute the query. It
prepares the query for execution.

The execution itself happens in the context of a multiplex, even if
there is only one query. This ensures that data that can be shared
between the queries, avoiding hitting the resources multiple times.

This changes the logging and instrumentation of those queries to share
the duration of the total execution. This ensures that we have
relevant duration information in the logs.

In practice, I don't think our frontends multiplex queries.

This is also the point at which the query is fully execution, which
means we can inspect the result to gather any errors that happened
during the execution of the query and expose that information in
metrics and logs.

This merges the 3 tracers that are supposed to provide information
into a single one that collects all of the information. This ensures
that we're always comparing apples to apples when we talk about
durations: the duration in the logs is also the duration we've used
for the apdex metric.

For https://gitlab.com/gitlab-org/gitlab/-/issues/345263