Work item hierarchy change system notes use full references
When a parent belongs to a different namespace than the
child. The system note about adding as a parent and as a
child need to use the full reference of the work items as
otherwise it will be interpreted as the IID of the same
project or group where the note is created.

Changelog: fixed