Remove unused issue card weight filtering functionality
The weight widget on an issue card used to be clickable to filter the
board on issues that have that weight, but this functionality has been
broken for years. This MR removes the broken functionality and aligns
the weight widget with the milestone, iteration and due date widget
styles on an issue card.

Changelog: fixed