Cleanup index terraform module archive FF
In order to be able to display documentation data on the UI page of a terraform module, we had to extract them from the uploaded archive. The extraction & parsing logic are done in a background worker, which was behind a feature flag. The flag was rolled out on production and things are stable. In this MR, we cleanup the feature flag from the codebase and release the feature.

Changelog: added