Add an error rate SLI for graphql queries
This adds a counter for all GraphQL queries. It increments the `ops`
rate for all queries exectuted and it increments the error rate in
case there was an exception or if the result contained errors.

This means that invalid queries sent to us will also result in an
error. For this reason, we need to make sure that we only include
queries that we know of in our SLIs. We can distinguish these in
metrics using the `endpoint_id` label. We'll only populate that label
for queries from our own application. All other queries will have
`graphql:unknown` as the `endpoint_id`.

For https://gitlab.com/gitlab-org/gitlab/-/issues/345263