Improve ProtectedRefAccess#type logic
Prior to this change we only checked if user or group was present. This
is not sufficient to ensure the correct type is returned. If a
ProtectedRefAccess level has a user_id or group_id assigned and the user
or group is not loaded (or doesn't exist) then :role would be returned
incorrectly.

This change checks both user_id and user, and group_id and group when
inferring the type.