Add worker to aggregate last 30-day catalog resource usage data
Introduces a new service and worker that aggregates the last
30-day component usage counts for all catalog resources. The
service is set up with a rolling 30-day window and utilizes the
Usages::Aggregator class which aggregates the usage data in
time-boxed batches.

The worker is scheduled every 4 minutes. On each run, the
aggregator resumes processing the usage data from where it last
left off. The usage counts are updated in the
catalog_resources.last_30_day_usage_count column. Each count
is expected to be updated only once per day.

This MR also adds database indexes to help optimize the queries
executed in the batched aggregation process.

Changelog: added