Ensure Workhorse log writers are closed to avoid Goroutine leaks
We've seen in production the number of Goroutines in Workhorse exceed
4000. The pprof output suggests that many of these came from the
logrus Goroutines lingering.

When Workhorse forks `gitlab-zip-cat` and `gitlab-zip-metadata` it
attempts to log all stderr via a logrus Writer. logrus launches a
Goroutine to listen from a pipe. We should ensure this Writer is
closed so that the Goroutines shut down as soon as possible.

Changelog: fixed