Ensure project can always be called on ProtectedRefAccess objects
In `ProtectedBranchAccess`, `ProtectedTagAccess`, and
`ProtectedBranch::UnprotectAccessLevel` we delegate project to the
protected ref e.g.

```ruby
delegate :project, to: :protected_branch
delegate :project, to: :protected_tag
```

If the `protected_branch` or `protected_tag` is nil a `NoMethodError`
will be raised when calling `#project`.

This change adds `allow_nil: true` to the delegations and covers the
issue in a spec.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/461213+