Implement AR validations for ProtectedRefAccess
Prior to this change we did not validate that the user or group actually
existed when creating a ProtectedRefAccess object.

We have foreign_key constraints but this means we relied on the database
layer to raise exceptions instead of returning user friendly errors.

Also refactors how the validations are structured so it's simpler to see
what validations are ignored during importing.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/461213

Changelog: changed
EE: true