Use LocalStorage for frequently used emoji
Previously, a cookie was used for this but (a) there is no need to use a
cookie for communicating with the backend here, all the logic is in the
frontend and (b) cookies expire. Prior to this MR, a user's frequently
used emojis list would periodically get cleared out. This change will
mean that a user's frequently used emoji list persists until such time
as the user manually clears their local storage.

Changelog: fixed