Change catalog resource last 30-day count updated_at
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155001,
we introduced a new aggregator worker that collects the last 30
day usage count for each catalog resource daily. It was observed
that it reprocesses the data more often than necessary due to
new catalog resources added throughout the day. This is because
the stop condition AggregateLast30DayUsageService#done_processing?
checks if MIN(last_30_day_usage_count_updated_at) >= today.to_time.
And when a new catalog resource is added, it defaults to 1970-01-01,
so it's seen as another unprocessed target.

This MR corrects this situation by updating the default to NOW(),
so that new catalog resources are added in an "already processed"
state.

Changelog: other