Fix segmented vulnerability export logic
While creating the export parts, we are iterating over all the
`vulnerability_reads` records ordered by `traversal_ids` and
`vulnerability_id`. This means a higher value can be set as
`start_id` than `end_id`.

For example for the the following tuples;

('{9970}', 10), ('{9970}', 11), ('{9970, 9972}', 1)

The `start_id` will be 10 and the `end_id` will be 1 as these records
are first sorted by `traversal_ids`.

For this reason, we need to compare the records by both `traversal_ids`
and `vulnerability_id` while iterating over the records of export part.