E2E: Add Repeaters to harden verify ide specs
During https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154802
we do a large scale refactor of the WebIDE E2E tests.
During monitoring of the changes it was observed in jobs such as
https://gitlab.com/gitlab-org/gitlab/-/jobs/7108653064 and
https://gitlab.com/gitlab-org/gitlab/-/jobs/7110957043
that with a slow network the check for has_message may be flaky.
I suspect this was previously masked by the slower duration of the
suite. We should allow for the test to wait longer to verify the
failure message.