Fix orphaned work item parent links
When moving an issue we do the following:
1. We close the initial issue `A` on the old project.
2. We create a new issue `B` on the new project.
3. We change the `epic_issues` relationship and update the
  `epic_issues.issue_id` to the ID of issue `B`.

We missed to sync step 3 to `work_item_parent_links`, and in addition to
that,  to update the `namespace_id` on `work_item_parent_links`.

`epic_issues` is still the SSoT and contains all epic<>issues
relationships we need to sync. However, we have orphaned
`work_item_parent_links` with the wrong `work_item_id` and
`namespace_id`.

This is what we need to fix as part of the migration. Since the wrong
syncing logic was never shipped to self managed instances, we can run
the migration only on SaaS.

Changelog: fixed
EE: true