Add subscription group creation endpoints
We need to be able to create a group in the GitLab purchase flow, before
the customer is sent to CustomersDot for the purchase.

Add two new actions to the Subscriptions::GroupsController to facilitate
the creation of a new group for subscription purchase.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/468097
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/156854