Fix AI review to not include meta lines as part of diff hunks
We previously include diff lines that are meta lines (lines with
type as `new-nonewline`, `old-nonewline`, etc) as part of diff
hunks we remove. It results to our worker creating a draft note
on that line which doesn't work (we don't allow commenting on
those lines even in the UI).

To fix this, we exclude them from the diff hunk we review so the
draft note is created at the correct line.