Do not memoize Vulnerabilities::Export#export_service method
The `VulnerabilityExports::ExportService#exporter` method is memoized,
which means, it returns the same `exporter` for all the calls and that
same exporter uses the "initial enumerator" value passed into it when
we call the `export_segment` method multiple times.

For this reason, only the first call of the `export_segment` method
generates the CSV file content for the export part and the subsequent
calls do not generate content as the initially given enumerator is
already completed and doesn't yield anything.