Validate provided group parameters
This change validates the provided params when attempting to update
a group. This ensures that members of a custom role that have the
`admin_runners` permission enabled can only update a single param.
This custom permission is currently gated behind a
[feature flag](https://gitlab.com/gitlab-org/gitlab/-/issues/461448)
and currently Owners have both the `admin_group` and `admin_runner`
permission.

EE: true
Changelog: fixed