Only enumerate commits in pre-receive check if push came from Web
In GitLab 17.0,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150466
introduced a push rule check that attempts to validate that a signed
commit doesn't have a customized author. This check requires calling
`ListAllCommitsRequest` to enumerate all the commits in the
quarantined repository.

This merge request introduced `Gitlab::Checks::CommitChecks` that
enumerates over all the commits behind the
`skip_committer_email_check` feature flag. However, regardless of the
state of this feature flag, `ListAllCommitsRequest` is still called.
This RPC can timeout for pushes with large number of commits,
resulting in a 500 error in the `/api/v4/internal/allowed` endpoint
and ultimately causing pushes to fail.

As it stands, the current feature only checks commits that were:

1. Signed by GitLab itself
2. Pushed via a UI change

We can short-circuit the commits enumeration in the common case by
checking whether the push came from the UI.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/467556

Changelog: fixed