Do not show cluster image scanning vulnerabilities on development tab
This MR makes sure that when no `reportTypes` or `scanner` are set,
the `reportTypes` defaults are set for the development tab.

This also properly tests the graphql filter functions in the
vulnerability_report_tabs component. It also simplifies the case where
the feature flag `containerScanningForRegistryFlag` is enabled by
making it the default case and removing unnecessary tests when the
feature flag is disabled.

Changelog: fixed
EE: true