Document read-only transaction for registry
If the registry fails to apply schema migrations or
to run online garbage collection with an error related to
not being able to execute a statement in a read-only
transaction, a setting may be preventing this.

This commit adds a note to the troubleshooting section with
the errors and how to disable the related setting
`default_transaction_read_only`.